Worked on integrating filters to Swipe and Match

+++ lib/feature/home/preferences/data/state/search_preferences_notifier.dart
@@ -16,7 +16,7 @@ abstract class SearchPreferencesNotifier {
     bool? hasBio,
     int? minAge,
     int? maxAge,
-    double? distance,
+    int? distance,
     bool? outreach,
     List<String>? interests,
     String? country,
