+++ lib/feature/home/tabs/explore/data/state/advanced_search_notifier.dart
@@ -97,7 +97,7 @@ class AdvancedSearchNotifier extends ChangeNotifier
     bool? hasBio,
     int? minAge,
     int? maxAge,
-    double? distance,
+    int? distance,
     bool? outreach,
     List<String>? interests,
     String? country,
